+++ 1.8.9/sources/compat/src/api/java/net/teamfruit/bnnwidget/compat/Compat.java
@@ -91,14 +91,14 @@ public class Compat {
 
 		@Override
 		public int drawStringWithShadow(@Nullable final String str, final float x, final float y, final int color) {
-			return drawStringWithShadowCompat(str, (int) x, (int) y, color);
+			return drawStringWithShadowCompat(str, x, y, color);
 		}
 
 		protected abstract int drawStringWithShadowCompat(@Nullable final String str, final float x, final float y, final int color);
 
 		@Override
 		public int drawString(@Nullable final String str, final float x, final float y, final int color, final boolean shadow) {
-			return drawStringCompat(str, (int) x, (int) y, color, shadow);
+			return drawStringCompat(str, x, y, color, shadow);
 		}
 
 		protected abstract int drawStringCompat(@Nullable final String str, final float x, final float y, final int color, final boolean shadow);
